requirements.txt
fastmcp
pandas
scikit-learn
pyngrok
python-dotenv
